refactor: move encoding struct into independent file
